Add files via upload
The project focused on Exploratory Data Analysis of a COVID-19 Clinical Trials dataset, employing Python libraries like Pandas, Matplotlib, and Seaborn. It aimed to dissect the landscape and progress of various studies centered around COVID-19, through a comprehensive dataset of 5,783 records. The analysis included the examination of study statuses, types of interventions, and the counts of outcome measures. Additionally, the study delved into the temporal distribution of trials, pinpointing trends in research focal points. Insights into the global response to the pandemic were gleaned, showcasing the foremost conditions being researched and the geographical spread of these studies. This endeavor illuminated the multifaceted response to COVID-19 and refined analytical capabilities in handling and interpreting complex datasets.

-- Utilized Python's Pandas library to manipulate and analyze a dataset containing over 5,783 COVID-19 clinical trials, providing a foundation for comprehensive data assessment.
-- Employed Matplotlib and Seaborn for visual data analysis, creating a series of plots to visualize the distribution of clinical trials' statuses and intervention types.
-- Identified key trends in research focus over time through temporal data analysis, highlighting shifts in the global research community's priorities in response to the pandemic.
-- Analyzed the frequency of various outcome measures within the trials, offering insights into common endpoints and objectives in COVID-19 research.
-- Explored the geographical distribution of studies, revealing top countries contributing to COVID-19 research, thereby underscoring the global nature of the scientific response.
-- Conducted a detailed examination of the top conditions being studied in relation to COVID-19, shedding light on the diverse areas of interest within the research community.